+++ maya_mcp_ui.py
@@ -28,6 +28,15 @@ WORKSPACE_NAME = "mcpListenerPanel"
 
 
 def _pid_alive(pid):
+    import sys
+    if sys.platform == "win32":
+        import ctypes
+        kernel32 = ctypes.windll.kernel32
+        handle = kernel32.OpenProcess(0x100000, False, pid)
+        if handle:
+            kernel32.CloseHandle(handle)
+            return True
+        return False
     try:
         os.kill(pid, 0)
         return True
